+++ bpf/btf/btf_helpers.h
@@ -79,4 +79,5 @@ const char *btf_kind_name(int kind);
 
 int btf_has_ksym(const char *ksym);
 int btf_has_kfunc(const char *kfunc);
+int btf_has_struct(const char *sname);
 int btf_has_decl_tag(const char *ksym);