+++ bpf/btf/btf_ksym.c
@@ -22,6 +22,10 @@ int main(int argc, char **argv)
 #pragma message "defined TEST_BTF_KIND_FUNC"
 	result = btf_has_kfunc(ksym_name);
 	s_type = "FUNC";
+#elif defined(TEST_BTF_KIND_STRUCT)
+#pragma message "defined TEST_BTF_KIND_STRUCT"
+	result = btf_has_struct(ksym_name);
+	s_type = "STRUCT";
 #elif defined(TEST_BTF_KIND_DECL_TAG)
 #pragma message "defined TEST_BTF_KIND_DECL_TAG"
 	result = btf_has_decl_tag(ksym_name);