+++ interrupt/cli.c
@@ -1,0 +1,6 @@
+int main(void)
+{
+	/* trigger X86_TRAP_GP */
+	asm volatile("cli");
+	return 0;
+}