+++ ipc/socket/iovec/iovec_helpers.h
@@ -1,3 +1,5 @@
+// SPDX-License-Identifier: (LGPL-2.1 OR BSD-2-Clause)
+/* Copyright (c) 2025 Rong Tao */
 #pragma once
 
 #define UNSOCKET_PATH   "/tmp/fd-pass.socket"