+++ ipc/socket/iovec/socketpair.c
@@ -7,7 +7,7 @@
 #include <string.h>
 #include <sys/socket.h>
 
-#include "common.h"
+#include "iovec_helpers.h"
 
 
 int main(int argc, const char *argv[])