+++ ipc/socket/iovec/unix-server.c
@@ -8,7 +8,7 @@
 #include <sys/wait.h>
 #include <sys/socket.h>
 
-#include "common.h"
+#include "iovec_helpers.h"
 
 int main(int argc, char *argv[])
 {