+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to
 
 #### Breaking Changes
 #### Added
+- bpf: add pure userspace code tcpdump.c;
 #### Changed
 #### Deprecated
 #### Removed