+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to
 #### Breaking Changes
 #### Added
 - bpf: add pure userspace code tcpdump.c;
+- kprobe: add README.md;
 #### Changed
 #### Deprecated
 #### Removed